# target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Repair Tips API
  description: API Backend para Sistema de Consulta de Dicas de Conserto de Equipamentos
  version: 1.0.0
  contact:
    name: API Support
    url: https://example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api
    description: Development Server
  - url: https://api.example.com/api
    description: Production Server
...
paths:
...
  /equipamentos:
    get:
      tags:
        - Equipamentos
      summary: Listar todos os equipamentos
      operationId: getEquipamentos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de equipamentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEquipamentoResponseDTO'
    post:
      tags:
        - Equipamentos
      summary: Criar novo equipamento
      operationId: createEquipamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipamentoRequestDTO'
      responses:
        '201':
          description: Equipamento criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipamentoResponseDTO'
  /equipamentos/categoria/{categoria}:
    get:
      tags:
        - Equipamentos
      summary: Buscar equipamentos por categoria
      operationId: getEquipamentosByCategoria
      parameters:
        - name: categoria
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/CategoriaEquipamento'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Equipamentos encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEquipamentoResponseDTO'
  /equipamentos/tipo/{tipo}:
    get:
      tags:
        - Equipamentos
      summary: Buscar equipamentos por tipo
      operationId: getEquipamentosByTipo
      parameters:
        - name: tipo
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Equipamentos encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEquipamentoResponseDTO'
  /equipamentos/fabricante/{fabricanteId}:
    get:
      tags:
        - Equipamentos
      summary: Buscar equipamentos por fabricante
      operationId: getEquipamentosByFabricante
      parameters:
        - name: fabricanteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Equipamentos encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEquipamentoResponseDTO'
  /equipamentos/{id}:
    get:
      tags:
        - Equipamentos
      summary: Obter equipamento por ID
      operationId: getEquipamentoById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Equipamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipamentoResponseDTO'
        '404':
          description: Equipamento não encontrado
    put:
      tags:
        - Equipamentos
      summary: Atualizar equipamento
      operationId: updateEquipamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipamentoRequestDTO'
      responses:
        '200':
          description: Equipamento atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipamentoResponseDTO'
    delete:
      tags:
        - Equipamentos
      summary: Deletar equipamento
      operationId: deleteEquipamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Equipamento deletado com sucesso
  /equipamentos/modelo/{modelo}:
    get:
      tags:
        - Equipamentos
      summary: Buscar equipamentos por modelo
      operationId: searchEquipamentosByModelo
      parameters:
        - name: modelo
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Equipamentos encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEquipamentoResponseDTO'
...
components:
  schemas:
    CategoriaEquipamento:
      type: string
      enum:
        - IMPRESSORA
        - SCANNER
        - COPIADORA
        - MULTIFUNCIONAL
        - FAX
        - PROJETOR
        - MONITOR
        - COMPUTADOR
        - SERVIDOR
        - IMPRESSORA_3D
        - OUTRO
      description: Categoria do equipamento
...
    EquipamentoRequestDTO:
      type: object
      required:
        - modelo
        - categoria
        - tipo
        - fabricanteId
      properties:
        modelo:
          type: string
          minLength: 1
          maxLength: 100
          description: Modelo do equipamento
        categoria:
          $ref: '#/components/schemas/CategoriaEquipamento'
        tipo:
          type: string
          minLength: 1
          maxLength: 50
          description: Tipo do equipamento
        fabricanteId:
          type: integer
          format: int64
          description: ID do fabricante
    EquipamentoResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        modelo:
          type: string
        categoria:
          $ref: '#/components/schemas/CategoriaEquipamento'
        tipo:
          type: string
        fabricanteId:
          type: integer
          format: int64
...
    PageEquipamentoResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EquipamentoResponseDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
...