# target/generated-sources/openapi/src/main/java/com/repairsystem/api/EquipamentosApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.repairsystem.api;

import com.repairsystem.dto.CategoriaEquipamento;
import com.repairsystem.dto.EquipamentoRequestDTO;
import com.repairsystem.dto.EquipamentoResponseDTO;
import com.repairsystem.dto.PageEquipamentoResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-23T23:19:45.937138696-03:00[America/Sao_Paulo]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "Equipamentos", description = "Operações com equipamentos")
public interface EquipamentosApi {

    default EquipamentosApiDelegate getDelegate() {
        return new EquipamentosApiDelegate() {};
    }

    /**
     * POST /equipamentos : Criar novo equipamento
     *
     * @param equipamentoRequestDTO  (required)
     * @return Equipamento criado com sucesso (status code 201)
     */
    @Operation(
        operationId = "createEquipamento",
        summary = "Criar novo equipamento",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Equipamento criado com sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = EquipamentoResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/equipamentos",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<EquipamentoResponseDTO> createEquipamento(
        @Parameter(name = "EquipamentoRequestDTO", description = "", required = true) @Valid @RequestBody EquipamentoRequestDTO equipamentoRequestDTO
    ) {
        return getDelegate().createEquipamento(equipamentoRequestDTO);
    }


    /**
     * DELETE /equipamentos/{id} : Deletar equipamento
     *
     * @param id  (required)
     * @return Equipamento deletado com sucesso (status code 204)
     */
    @Operation(
        operationId = "deleteEquipamento",
        summary = "Deletar equipamento",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Equipamento deletado com sucesso")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/equipamentos/{id}"
    )
    
    default ResponseEntity<Void> deleteEquipamento(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return getDelegate().deleteEquipamento(id);
    }


    /**
     * GET /equipamentos/{id} : Obter equipamento por ID
     *
     * @param id  (required)
     * @return Equipamento encontrado (status code 200)
     *         or Equipamento não encontrado (status code 404)
     */
    @Operation(
        operationId = "getEquipamentoById",
        summary = "Obter equipamento por ID",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Equipamento encontrado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = EquipamentoResponseDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Equipamento não encontrado")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/equipamentos/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<EquipamentoResponseDTO> getEquipamentoById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return getDelegate().getEquipamentoById(id);
    }


    /**
     * GET /equipamentos : Listar todos os equipamentos
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Lista de equipamentos (status code 200)
     */
    @Operation(
        operationId = "getEquipamentos",
        summary = "Listar todos os equipamentos",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de equipamentos", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageEquipamentoResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/equipamentos",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PageEquipamentoResponseDTO> getEquipamentos(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Optional<Integer> page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Optional<Integer> size
    ) {
        return getDelegate().getEquipamentos(page, size);
    }


    /**
     * GET /equipamentos/categoria/{categoria} : Buscar equipamentos por categoria
     *
     * @param categoria  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Equipamentos encontrados (status code 200)
     */
    @Operation(
        operationId = "getEquipamentosByCategoria",
        summary = "Buscar equipamentos por categoria",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Equipamentos encontrados", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageEquipamentoResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/equipamentos/categoria/{categoria}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PageEquipamentoResponseDTO> getEquipamentosByCategoria(
        @Parameter(name = "categoria", description = "", required = true, in = ParameterIn.PATH) @PathVariable("categoria") CategoriaEquipamento categoria,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Optional<Integer> page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Optional<Integer> size
    ) {
        return getDelegate().getEquipamentosByCategoria(categoria, page, size);
    }


    /**
     * GET /equipamentos/fabricante/{fabricanteId} : Buscar equipamentos por fabricante
     *
     * @param fabricanteId  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Equipamentos encontrados (status code 200)
     */
    @Operation(
        operationId = "getEquipamentosByFabricante",
        summary = "Buscar equipamentos por fabricante",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Equipamentos encontrados", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageEquipamentoResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/equipamentos/fabricante/{fabricanteId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PageEquipamentoResponseDTO> getEquipamentosByFabricante(
        @Parameter(name = "fabricanteId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("fabricanteId") Long fabricanteId,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Optional<Integer> page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Optional<Integer> size
    ) {
        return getDelegate().getEquipamentosByFabricante(fabricanteId, page, size);
    }


    /**
     * GET /equipamentos/tipo/{tipo} : Buscar equipamentos por tipo
     *
     * @param tipo  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Equipamentos encontrados (status code 200)
     */
    @Operation(
        operationId = "getEquipamentosByTipo",
        summary = "Buscar equipamentos por tipo",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Equipamentos encontrados", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageEquipamentoResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/equipamentos/tipo/{tipo}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PageEquipamentoResponseDTO> getEquipamentosByTipo(
        @Parameter(name = "tipo", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tipo") String tipo,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Optional<Integer> page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Optional<Integer> size
    ) {
        return getDelegate().getEquipamentosByTipo(tipo, page, size);
    }


    /**
     * GET /equipamentos/modelo/{modelo} : Buscar equipamentos por modelo
     *
     * @param modelo  (required)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Equipamentos encontrados (status code 200)
     */
    @Operation(
        operationId = "searchEquipamentosByModelo",
        summary = "Buscar equipamentos por modelo",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Equipamentos encontrados", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageEquipamentoResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/equipamentos/modelo/{modelo}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PageEquipamentoResponseDTO> searchEquipamentosByModelo(
        @Parameter(name = "modelo", description = "", required = true, in = ParameterIn.PATH) @PathVariable("modelo") String modelo,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Optional<Integer> page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Optional<Integer> size
    ) {
        return getDelegate().searchEquipamentosByModelo(modelo, page, size);
    }


    /**
     * PUT /equipamentos/{id} : Atualizar equipamento
     *
     * @param id  (required)
     * @param equipamentoRequestDTO  (required)
     * @return Equipamento atualizado (status code 200)
     */
    @Operation(
        operationId = "updateEquipamento",
        summary = "Atualizar equipamento",
        tags = { "Equipamentos" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Equipamento atualizado", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = EquipamentoResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/equipamentos/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<EquipamentoResponseDTO> updateEquipamento(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "EquipamentoRequestDTO", description = "", required = true) @Valid @RequestBody EquipamentoRequestDTO equipamentoRequestDTO
    ) {
        return getDelegate().updateEquipamento(id, equipamentoRequestDTO);
    }

}